src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/custom/lifecycle-created-createElement-recursion-expected.txt

CONSOLE ERROR: line 8: Uncaught TypeError: document.registerElement is not a function
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
